refactor: remove status from update command, resolve SpotBugs warnings
Status is now managed exclusively via /start and /close endpoints,
preventing bypass of the state machine through PUT. Replaced
@SuppressFBWarnings with defensive List.copyOf() in record compact
constructors. Added SpotBugs exclude filter for generated JPA metamodel
classes. Moved StaleVersionException handler to GlobalExceptionHandler.
