+++ backend/repair-orders.http
@@ -58,8 +58,7 @@ Content-Type: application/json
 If-Match: 0
 
 {
-  "description": "Engine diagnostic — completed",
-  "status": "COMPLETED",
+  "description": "Engine diagnostic — updated notes",
   "vehicleMake": "Ford",
   "vehicleModel": "F-150",
   "vehicleYear": 2023,
@@ -120,7 +119,6 @@ If-Match: 999
 
 {
   "description": "Should fail",
-  "status": "COMPLETED",
   "vehicleMake": "Toyota",
   "vehicleModel": "Camry",
   "vehicleYear": 2021
