+++ backend/src/main/java/com/interview/controller/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.interview.controller;
 
+import com.interview.exception.StaleVersionException;
 import jakarta.validation.ConstraintViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ProblemDetail;
@@ -27,5 +28,12 @@ public class GlobalExceptionHandler {
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ProblemDetail handleIllegalArgument(IllegalArgumentException ex) {
         return ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, ex.getMessage());
+    }
+
+    @ExceptionHandler(StaleVersionException.class)
+    @ResponseStatus(HttpStatus.PRECONDITION_FAILED)
+    public ProblemDetail handleStaleVersion(StaleVersionException ex) {
+        return ProblemDetail.forStatusAndDetail(HttpStatus.PRECONDITION_FAILED,
+            ex.getMessage());
     }
 }
