+++ backend/src/main/java/com/interview/dto/PageDto.java
@@ -1,10 +1,7 @@
 package com.interview.dto;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.util.List;
 
-@SuppressFBWarnings(value = "EI_EXPOSE_REP",
-    justification = "Records are immutable; list contents are not mutated")
 public record PageDto<T>(
     List<T> content,
     int page,
@@ -12,4 +9,9 @@ public record PageDto<T>(
     long totalElements,
     int totalPages,
     boolean last
-) {}
+) {
+
+  public PageDto {
+    content = content == null ? List.of() : List.copyOf(content);
+  }
+}
