+++ backend/src/main/java/com/interview/dto/UpdateRepairOrderCommand.java
@@ -1,13 +1,12 @@
 package com.interview.dto;
 
-import com.interview.model.RepairOrderStatus;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+// Status is managed exclusively via /start and /close endpoints — not updatable here.
 public record UpdateRepairOrderCommand(
     @NotBlank String description,
-    @NotNull RepairOrderStatus status,
     @NotBlank String vehicleMake,
     @NotBlank String vehicleModel,
     @NotNull @Positive Integer vehicleYear,
